refactor: Refactor jios card list search params
