+++ frontend/src/pages/dashboard/jios/JioCardList.tsx
@@ -10,14 +10,13 @@ import EmptyContent from '../../../components/EmptyContent';
 import EmptyJiosContent from './EmptyJiosContent';
 import { PATH_DASHBOARD } from 'src/routes/paths';
 import { Link } from 'react-router-dom';
+import { setDateTime } from 'src/utils/formatTime';
 
-interface JioCardListProps {
-  searchParams: ListJiosArgs;
-}
-
-export default function JioCardList({ searchParams }: JioCardListProps) {
+export default function JioCardList() {
   const dispatch = useDispatch();
   const version = useVersion();
+  const jioFormValues = useSelector((state) => state.jioFormValues.data);
+
   const { data, error, loading } = useSelector((state) => state.jios);
   const displayedData = React.useMemo(() => {
     if (error) {
@@ -51,8 +50,40 @@ export default function JioCardList({ searchParams }: JioCardListProps) {
   }, [data, error, loading]);
 
   useEffect(() => {
+    // TODO: move to utils?
+    const getStartDateTimeString = () => {
+      if (jioFormValues?.date && jioFormValues?.startDateTime) {
+        return setDateTime(jioFormValues.date, jioFormValues.startDateTime).toISOString();
+      }
+
+      return undefined;
+    };
+    const getEndDateTimeString = () => {
+      if (jioFormValues?.date && jioFormValues?.endDateTime) {
+        return setDateTime(jioFormValues.date, jioFormValues.endDateTime).toISOString();
+      }
+
+      return undefined;
+    };
+    const searchParams: ListJiosArgs = {
+      type: jioFormValues?.type,
+      numPasses: jioFormValues?.numPasses,
+      gymId: jioFormValues?.gymId,
+      startDateTime: getStartDateTimeString(),
+      endDateTime: getEndDateTimeString(),
+    };
+
     dispatch(listJios(searchParams));
-  }, [version, dispatch, searchParams]);
+  }, [
+    version,
+    dispatch,
+    jioFormValues?.type,
+    jioFormValues?.numPasses,
+    jioFormValues?.gymId,
+    jioFormValues?.date,
+    jioFormValues?.startDateTime,
+    jioFormValues?.endDateTime,
+  ]);
 
   return <Grid container>{displayedData}</Grid>;
 }
